+++ CHANGELOG.md
@@ -88,6 +88,7 @@ Put the changelog BELOW the dashes. ANYTHING ABOVE IS IGNORED.
 - Calcite now generates alongside Sulfur and Cinnabar in the Sulfur Caves biome.
   - Added a config option to toggle Calcite generation.
 - Shrubs can no longer generate on Sand blocks in the Oasis biome.
+- Added the Icicle to the `minecraft:speleothems` Block Tag.
 - Added the `wilderwild:sulfur_caves_calcite_enabled` Material Condition.
 - Renamed the `wilderwild:beta_beach_condition_source` Material Condition to `wilderwild:beta_beach_enabled`.
 - Renamed the `wilderwild:snow_under_mountain_condition_source` Material Condition to `wilderwild:snow_under_mountain_enabled`.